requirements.txt
PyMySQL>=1.1.0
dnspython>=2.6.0
